Refactor length of stay calculation and update schema type
- Changed the calculation of length of stay in `build_record` to return an integer number of days instead of a rounded float.
- Updated the `DischargeRecord` schema to reflect the change in type for `length_of_stay_days` from float to int.
- Adjusted logging format to display length of stay as an integer.

+++ src/ingestion/schema.py
@@ -13,7 +13,7 @@ class DischargeRecord:
     encounter_id: str
     admit_ts: str
     discharge_ts: str
-    length_of_stay_days: float
+    length_of_stay_days: int
     age_at_discharge: int
 
     admission_reason: str                     # Encounter.reasonCode — 100% populated, confirmed
